Jun 30 Updates
Several Updates to improve the functionality of the Drop Down List
control.

Clicking outside the DDL control will close the dropped down list box.
When a DropDownList selection is reset (SelectedIndex = -1), the
selected text is reset to the placeholder's text.
Streamlined several bits of code.
Improved the Editor/Inspectors for the ListBox Control, DropDownList
Control and the ListItem prefab.